+++ card.txt
@@ -1,0 +1,52 @@
+heart,Ace,1
+heart,2,2
+heart,3,3
+heart,4,4
+heart,5,5
+heart,6,6
+heart,7,7
+heart,8,8
+heart,9,9
+heart,10,10
+heart,Jack,11
+heart,Queen,12
+heart,King,13
+diamond,Ace,1
+diamond,2,2
+diamond,3,3
+diamond,4,4
+diamond,5,5
+diamond,6,6
+diamond,7,7
+diamond,8,8
+diamond,9,9
+diamond,10,10
+diamond,Jack,11
+diamond,Queen,12
+diamond,King,13
+spade,Ace,1
+spade,2,2
+spade,3,3
+spade,4,4
+spade,5,5
+spade,6,6
+spade,7,7
+spade,8,8
+spade	9,9
+spade,10,10
+spade,Jack,11
+spade,Queen,12
+spade,King,13
+club,Ace,1
+club,2,2
+club,3,3
+club,4,4
+club,5,5
+club,6,6
+club,7,7
+club,8,8
+club,9,9
+club,10,10
+club,Jack,11
+club,Queen,12
+club,King,13
